fixed bugs with messaging/singlepostview

+++ src/components/DeleteUserPosts.js
@@ -1,16 +1,24 @@
 import React from "react";
+import { useNavigate } from "react-router-dom";
 import { deletePosts, retrievePosts } from "../api";
 import "./css/DeleteUserPosts.css";
 
 // This component is specific to deleting a post. We have it re-render the page after a post has been deleted without needing to refresh the page.
 
-const DeleteUserPosts = ({ element_id, setPosts }) => {
+const DeleteUserPosts = ({ element_id, setPosts, singlePostViewDelete }) => {
+  let navigate = useNavigate();
+  
   async function handleSubmit(event) {
     event.preventDefault();
     const token = localStorage.getItem("token");
-    await deletePosts(token, element_id);
-    const allPosts = await retrievePosts();
-    setPosts(allPosts.data.posts);
+    if (confirm("Do you want to delete this post?\nIf so, please click OK.\nOtherwise click Cancel.") === true) {
+      await deletePosts(token, element_id);
+      const allPosts = await retrievePosts();
+      setPosts(allPosts.data.posts);
+      if (singlePostViewDelete) {
+        navigate("/posts")
+      }
+    }
   }
 
   return (
